+++ variables.tf
@@ -252,19 +252,24 @@ variable "enable_acs" {
   default     = false
 }
 
-variable "acs_azuredisk_enabled" {
-  description = "Whether Azure Disks is enabled as an ACS backend"
-  type        = bool
-  default     = true
-}
-
-variable "acs_azuredisk_sku" {
-  description = "Azure Disk disk tier"
-  type        = string
-  default     = "Premium_LRS"
+variable "acs_config" {
+  description = "Azure Container Storage configuration settings"
+  type = object({
+    activecontrol                   = optional(bool, true)
+    storagepool_install_create      = optional(bool, false)
+    storagepool_disable_validation  = optional(bool, false)
+    storagepool_disable_active      = optional(bool, false)
+    storagepool_azuredisk_enabled   = optional(bool, true)
+    storagepool_azuredisk_sku       = optional(string, "Premium_LRS")
+    resources_num_hugepages         = optional(number, 512)
+    resources_ioengine_cpu          = optional(number, 1)
+    resources_ioengine_memory       = optional(string, "1Gi")
+    resources_ioengine_hugepages2mi = optional(string, "1Gi")
+  })
+  default = {}
 
   validation {
-    condition     = contains(["Premium_LRS", "Standard_LRS", "StandardSSD_LRS", "UltraSSD_LRS", "Premium_ZRS", "PremiumV2_LRS", "StandardSSD_ZRS"], var.acs_azuredisk_sku)
+    condition     = contains(["Premium_LRS", "Standard_LRS", "StandardSSD_LRS", "UltraSSD_LRS", "Premium_ZRS", "PremiumV2_LRS", "StandardSSD_ZRS"], var.acs_config.storagepool_azuredisk_sku)
     error_message = "Accepted values are 'Premium_LRS', 'Standard_LRS', 'StandardSSD_LRS', 'UltraSSD_LRS', 'Premium_ZRS', 'PremiumV2_LRS', or 'StandardSSD_ZRS'"
   }
 }